+++ releasenotes.md
@@ -27,6 +27,7 @@
 - Remove `[?]` syntax.
 - Change `int!` to `int?` syntax.
 - New `fault` declarations.
+- Enum associated values can reference the calling enum.
 
 ### Fixes
 - Fix address sanitizer to work on MachO targets (e.g. MacOS).